Improve code quality

+++ distance.h
@@ -1,10 +1,8 @@
 #ifndef KEYBORG_DISTANCE_H
 #define KEYBORG_DISTANCE_H
 
-#include <Arduino.h>
-
 void initDistance();
 
-uint16_t readDistance();
+float readDistance();
 
 #endif
